==> notebooks/01-jsg-SES_data_exploration.ipynb <==
import sys
print(f'Interpreter dir: {sys.executable}')
import os

if os.path.basename(os.getcwd()) == 'notebooks':
    os.chdir('../')
    
print(f'Working dir: {os.getcwd()}')
%load_ext autoreload
%autoreload 2

import pandas as pd

# Explore SES dataset

df1 = pd.read_csv("data/raw/SES-EFF13 (Flow)_15 Minute Flow_(m3hr).csv")
df1


df2 = pd.read_csv("data/raw/SES-KNT11 (Flow)_15 Minute Flow_(m3hr).csv")
df2

df_meter15 = pd.read_excel("data/raw/SES-Meter Data - 15mins.xlsx",sheet_name=None)

print(df_meter15.keys())
display(df_meter15['000F4252'])

df_meterd = pd.read_excel("data/raw/SES-Meter Data - Daily Reads.xlsx",sheet_name=None)
print(df_meterd.keys())
display(df_meterd['Additional Info'])
display(df_meterd['Meter 1'])

### Conclusion

SES datasets are incomplere: there is no link from the
