==> notebooks/06-jsg-smartMeter.ipynb <==
import sys
print(f'Interpreter dir: {sys.executable}')
import os

if os.path.basename(os.getcwd()) == 'notebooks':
    os.chdir('../')
    
print(f'Working dir: {os.getcwd()}')
%load_ext autoreload
%autoreload 2

import glob
from pathlib import Path
    
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', 100)

from waterhack.utils import find_col

# Load smart meter data

- Load smart meter attributes and acord data. Merge them.
- Then load all smart meter datasets
- Filter for extreme values of >100 and <-100
- Filter for the weird bug Louis mentiones
- Aggregated the consumptions by “CATEGORY ID” calculating the sum, mean, min, mas std

---

df_attrib = pd.read_excel("data/raw/AWSmartMeterAttributes.xlsx")
df_acord = pd.read_excel('data/raw/AcornData.xlsx')
df_type = pd.read_csv('data/raw/Acorn_Type_Reference_2020.csv')

display(df_attrib)
display(df_acord)
display(df_type)


df_attrib2 = pd.merge(df_attrib, df_acord, left_on='Unique Ref', right_on='Unique Reference')
df_attrib3 = pd.merge(df_attrib2, df_type, left_on='HouseholdAcornCategory', right_on = 'TYPE ID')
df_attrib3.drop(["Unique Reference","TYPE ID"], axis=1, inplace=True) # Drop duplicated cols
df_attrib3

# Apply a groupping function to each file, and concatenate at the end

file_list = list(Path('data/raw/SmartMeter/').rglob('*.txt'))
file_list

def process_file(f,df_attrib3):
    # Load file
    df_file = pd.read_csv(f, parse_dates=['DateTime'])
    
    ## Fix bug: WHen more than 1 record, according to Louis, sum and divide by 2
        # Aggregare by ref and datetime -> count -> fix -> rename cols
    df_c = df_file.groupby(['DateTime','UniqueReference']).agg({'Consumption':['sum','count']}).reset_index()
    df_c.columns = [f'{i}_{j}' if j != '' else f'{i}' for i,j in df_c.columns]
    # Count and fix
    print(f"\tFound {np.sum(df_c.Consumption_count > 1)} ({np.mean(df_c.Consumption_count > 1)}) rows with count > 1")
    df_c.loc[df_c.Consumption_count > 1,'Consumption_sum'] = df_c.loc[df_c.Consumption_count > 1,'Consumption_sum'] / 2
    
    # Drop and rename columns
    df_c = df_c.drop('Consumption_count',axis=1).rename(columns = {'Consumption_sum':'Consumption'})
    
    # Filter for strange bbehaviours if any
    idx_filter = (df_c.Consumption < 100) & (df_c.Consumption > -100 )
    print(f"\t{np.sum(~idx_filter)} rows ({np.mean(~idx_filter) })columns filtered out of (-100,100) ")
    df_c = df_c[idx_filter]
    
    # Merge with DMA info
    df_m = pd.merge(df_attrib3[['DMASHORTNAME','Unique Ref','CATEGORY ID']], df_c,left_on='Unique Ref',right_on='UniqueReference')
    
    # Aggregate
    df_agg = df_m.groupby(['DateTime','DMASHORTNAME','CATEGORY ID']).agg({'Consumption':['sum','mean','max','min','std'],})
    
    # Fix columns names
    df = df_agg.reset_index()
    df.columns = [f'{i}_{j}' if j != '' else f'{i}' for i,j in df.columns]
    
    return df

#TEST
df = process_file("data/raw/SmartMeter/SmartMeterJuly2020.txt", df_attrib3)

%%time
list_df = []
for i,f in enumerate(file_list):
    
    print(f"Processing file {f}, {i}/{len(file_list)}")
    df_file = process_file(f, df_attrib3)

    # Append
    list_df.append(df_file)

#df_all = pd.concat(list_df)
df_all = df_all.sort_values(['DateTime','DMASHORTNAME','CATEGORY ID'])
df_all

df_all.to_csv('data/processed/smart_meter_aggregated_category_V1.csv', index=False)

# Take hourly values from Smart meter aggregated data, and turn into daily and night volumes


def agg_DMA_Cat(df):
    
    df_filt = df[['DateTime','Consumption_sum']].copy()
    
    df_filt["Date"] = pd.to_datetime(df_filt["DateTime"]).dt.date
    df_filt["Time"] = pd.to_datetime(df_filt["DateTime"]).dt.time
    df_filt["Hour"] = pd.to_datetime(df_filt["DateTime"]).dt.hour

    # Daily volume
    df_Vd = df_filt.groupby("Date")['Consumption_sum'].mean()

    # Night volue
    df_Vn = df_filt[(df_filt["Hour"] >= 2) & (df_filt["Hour"] < 4)]
    df_Vn = df_Vn.groupby("Date").mean()

    # Merge both datasets
    df_solution = pd.merge(
        df_Vd, df_Vn, left_index=True, right_index=True, suffixes=("_Vd", "_Vn")
    )
    
    df_solution = df_solution.drop(["Hour"], axis=1)

    
    return df_solution

# TEST:
# selected_DMA = "NEWSEVMA"
# selected_id = 2
# df_filt = df[ (df['DMASHORTNAME'] == selected_DMA) & (df['CATEGORY ID'] == selected_id)]
# df_solution = agg_DMA_Cat(df_filt)
# df_solution

df = df_all

# Gorupby and apply
df_sol = df.groupby(['DMASHORTNAME','CATEGORY ID']).apply(agg_DMA_Cat).reset_index()

# Add holiday indicator
df_sol['Date'] = pd.to_datetime(df_sol.Date)
df_sol["is_holiday"] = np.where(df_sol.Date.dt.dayofweek >= 5, 1, 0)
df_sol

df_sol = df_sol.sort_values(['Date','DMASHORTNAME','CATEGORY ID'])
df_sol.to_csv('data/processed/smart_meter_night_day_V1.csv', index=False)
